requirements.txt
# Core FastAPI stack
fastapi==0.119.0
uvicorn==0.38.0
starlette==0.48.0
python-multipart==0.0.20
sse-starlette==3.0.2

# OpenAI & Agent tooling
openai==2.5.0
openai-agents==0.4.0
openai-chatkit==1.0.2

# Data validation & settings
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0

# HTTP clients
httpx==0.28.1
httpx-sse==0.4.3
httpcore==1.0.9
h11==0.16.0
requests==2.32.5
redis==5.0.4

# Configuration & environment
python-dotenv==1.1.1

# Schema & validation utilities
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.27.1

# Model Context Protocol
mcp==1.18.0

# Type checking & utilities
typing_extensions==4.15.0
typing-inspection==0.4.2
types-requests==2.32.4.20250913

# Utilities
PyJWT==2.10.1
tqdm==4.67.1
jiter==0.11.1
idna==3.11
sniffio==1.3.1
urllib3==2.5.0
distro==1.9.0
griffe==1.14.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
